+++ ewm-service/src/main/java/ru/practicum/compilation/service/impl/CompilationServiceAdminImpl.java
@@ -1,5 +1,8 @@
 package ru.practicum.compilation.service.impl;
 
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Service;
 import ru.practicum.compilation.exception.CompilationException;
 import ru.practicum.compilation.model.Compilation;
@@ -12,19 +15,17 @@ import ru.practicum.event.model.Event;
 import ru.practicum.event.repository.EventRepository;
 import ru.practicum.mappers.CompilationMapper;
 
+import javax.transaction.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
+@Transactional
+@RequiredArgsConstructor
+@FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
 public class CompilationServiceAdminImpl implements CompilationServiceAdmin {
-    private final CompilationRepository compilationRepository;
-    private final EventRepository eventRepository;
-
-    public CompilationServiceAdminImpl(CompilationRepository compilationRepository,
-                                       EventRepository eventRepository) {
-        this.compilationRepository = compilationRepository;
-        this.eventRepository = eventRepository;
-    }
+    CompilationRepository compilationRepository;
+    EventRepository eventRepository;
 
     @Override
     public CompilationDto addCompilation(NewCompilationDto newCompilationDto) {
@@ -36,15 +37,15 @@ public class CompilationServiceAdminImpl implements CompilationServiceAdmin {
 
     @Override
     public void deleteCompilationById(long complId) {
-        Compilation compilation = compilationRepository.findById(complId)
-                .orElseThrow(() -> new CompilationException("Подборка не найдена"));
+        if (!compilationRepository.existsById(complId)) {
+            throw new CompilationException("Подборка не найдена");
+        }
         compilationRepository.deleteById(complId);
     }
 
     @Override
     public void deleteEventFromCompilation(long complId, long eventId) {
-        Compilation compilation = compilationRepository.findById(complId)
-                .orElseThrow(() -> new CompilationException("Подборка не найдена"));
+        Compilation compilation = getCompilationOrThrowCompilationNotFound(complId);
         List<Event> eventList = compilation.getEvents().stream()
                 .filter(event -> event.getId() != eventId)
                 .collect(Collectors.toList());
@@ -54,8 +55,7 @@ public class CompilationServiceAdminImpl implements CompilationServiceAdmin {
 
     @Override
     public CompilationDto addEventIntoCompilation(long complId, long eventId) {
-        Compilation compilation = compilationRepository.findById(complId)
-                .orElseThrow(() -> new CompilationException("Подборка не найдена"));
+        Compilation compilation = getCompilationOrThrowCompilationNotFound(complId);
         Event event = eventRepository.findById(eventId)
                 .orElseThrow(() -> new EventNotFoundException(eventId));
         compilation.getEvents().add(event);
@@ -64,17 +64,20 @@ public class CompilationServiceAdminImpl implements CompilationServiceAdmin {
 
     @Override
     public CompilationDto unpinCompilationFromMainPage(long complId) {
-        Compilation compilation = compilationRepository.findById(complId)
-                .orElseThrow(() -> new CompilationException("Подборка не найдена"));
+        Compilation compilation = getCompilationOrThrowCompilationNotFound(complId);
         compilation.setPinned(false);
         return CompilationMapper.INSTANCE.toCompilationDto(compilationRepository.save(compilation));
     }
 
     @Override
     public CompilationDto pinCompilationOnMainPage(long complId) {
-        Compilation compilation = compilationRepository.findById(complId)
-                .orElseThrow(() -> new CompilationException("Подборка не найдена"));
+        Compilation compilation = getCompilationOrThrowCompilationNotFound(complId);
         compilation.setPinned(true);
         return CompilationMapper.INSTANCE.toCompilationDto(compilationRepository.save(compilation));
+    }
+
+    private Compilation getCompilationOrThrowCompilationNotFound(long complId) {
+        return compilationRepository.findById(complId)
+                .orElseThrow(() -> new CompilationException("Подборка не найдена"));
     }
 }
